+++ java/core/src/test/java/co/worklytics/psoxy/storage/impl/RecordBulkDataSanitizerImplTest.java
@@ -79,6 +79,7 @@ class RecordBulkDataSanitizerImplTest {
         Container.ForConfigService.class,
         MockModules.ForSecretStore.class,
         MockModules.ForHostEnvironment.class,
+        MockModules.ForResourceService.class,
         TestModules.ForProxyConstants.class,
     })
     public interface Container {