+++ java/core/src/test/java/co/worklytics/test/MockModules.java
@@ -224,5 +224,14 @@ public class MockModules {
             return mock(AsyncApiDataRequestHandler.class);
         }
     }
+
+    @Module
+    public interface ForResourceService {
+        @Provides
+        @Singleton
+        static ResourceService resourceService() {
+            return new co.worklytics.psoxy.gateway.impl.NoOpResourceService();
+        }
+    }
 }